refactor: 💡 Rename CookpiePrompt to CookiePromptModal
