+++ apps/webstore/src/app/app.component.html
@@ -1,4 +1,4 @@
 <webstore-navbar></webstore-navbar>
 <router-outlet></router-outlet>
 <webstore-footer></webstore-footer>
-<webstore-cookie-prompt></webstore-cookie-prompt>
+<webstore-cookie-prompt-modal></webstore-cookie-prompt-modal>
