+++ apps/webstore/src/app/pages/issues/rejected-cookies/rejected-cookies.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
-import { CookieStatus } from '../../../shared/components/cookie-prompt/cookie-prompt.constants';
+import { CookieStatus } from '../../../shared/components/modals/cookie-prompt-modal/cookie-prompt.constants';
 import { TermsOfUseModalComponent } from '../../../shared/components/modals/terms-of-use-modal/terms-of-use-modal.component';
 import {
   LocalStorageService,
