+++ apps/webstore/src/app/shared/components/modals/cookie-prompt-modal/cookie-prompt.constants.ts
