+++ apps/webstore/src/app/shared/services/local-storage/local-storage.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, OnDestroy } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
 import { LocaleType } from '../../../i18n';
-import { CookieStatus } from '../../components/cookie-prompt/cookie-prompt.constants';
+import { CookieStatus } from '../../components/modals/cookie-prompt-modal/cookie-prompt.constants';
 import { LocalStorageVars } from './local-storage.constants';
 
 interface ICache {
